--- a/Dockerfile
+++ b/Dockerfile
@@ -25,8 +25,9 @@ RUN pip install --no-cache-dir -r requirements.txt
 # Copy project
 COPY . .
 
-# Create non-root user
+# Create non-root user and data directory
 RUN useradd --create-home --shell /bin/bash app && \
+    mkdir -p /app/data && \
     chown -R app:app /app
 USER app