--- a/linting_env/bin/mypy
+++ b/linting_env/bin/mypy
@@ -1,0 +1,8 @@
+#!/home/devag/git/dev-assist/linting_env/bin/python3
+# -*- coding: utf-8 -*-
+import re
+import sys
+from mypy.__main__ import console_entry
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(console_entry())