--- a/linting_env/bin/pip3
+++ b/linting_env/bin/pip3
@@ -1,0 +1,8 @@
+#!/home/devag/git/dev-assist/linting_env/bin/python3
+# -*- coding: utf-8 -*-
+import re
+import sys
+from pip._internal.cli.main import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())