--- a/linting_env/bin/python3
+++ b/linting_env/bin/python3
@@ -1,0 +1,1 @@
+/usr/bin/python3